% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sierra_extended_pal.R
\name{sierra_ext_pal}
\alias{sierra_ext_pal}
\title{Sierra color palette}
\usage{
sierra_ext_pal(name = "spectral", n = 10, reverse = FALSE)
}
\arguments{
\item{name}{name of palette}

\item{n}{number of colors}

\item{reverse}{reverse color palette}
}
\description{
Sierra's secondary color palette for discrete plotting
}
\examples{
library(scales)
library(tidyverse)
#library(sierra.plot)

show_col(sierra_ext_pal(n=12)(12)) # Default palette: spectral
show_col(sierra_ext_pal(name = 'greens', n=4)(4))
show_col(sierra_ext_pal(name = 'long', n=40)(40))
}
\concept{color sierra}
