% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sierra_simple.R
\name{theme_sierra_simple}
\alias{theme_sierra_simple}
\title{Sierra College Inspired Theme}
\usage{
theme_sierra_simple(base_size = 11, base_family = "serif")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
For GGPlotting in R
\href{https://www.sierracollege.edu/assets/docs/marketing/branding-logos/sierra\%20brand\%20guidelines\%201-17.pdf}{Sierra College Branding}
}
\examples{
library(tidyverse)
library(sierra.plot)

# Simple Plot, no color, fill, or groups to show base theme
ggplot(mtcars, aes(x= cyl, y= mpg)) +
  geom_point() +
  theme_sierra_simple()

# Bar plot with fill, using Sierra colors
ggplot(mtcars, aes(y= fct_rev(factor(cyl)))) +
  geom_bar(stat = "count") +
  theme_sierra_simple()

# Line plot with Sierra colored lines
ggplot(mtcars, aes(x= wt, y = mpg, color = factor(cyl), group = cyl)) +
  geom_line(size = 2) +
  scale_color_sierra(name = 'greens', n=4, reverse= TRUE) +
  theme_sierra_simple()
}
\seealso{
Other theme: 
\code{\link{theme_sierra}()}
}
\concept{theme}
